# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hub Marketplace API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: products
    description: Product management
  - name: sync
    description: Marketplace synchronization
  - name: questions
    description: Mercado Livre customer questions
  - name: ml-oauth
    description: Mercado Livre OAuth
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products for user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted

  /dashboard:
    get:
      operationId: getDashboard
      tags: [products]
      summary: Get dashboard stats
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /sync:
    post:
      operationId: syncPlatform
      tags: [sync]
      summary: Sync with a marketplace platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"

  /sync/logs:
    get:
      operationId: getSyncLogs
      tags: [sync]
      summary: Get sync history
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SyncLog"

  /questions/webhook:
    post:
      operationId: receiveQuestion
      tags: [questions]
      summary: Receive question from n8n webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/N8nQuestionPayload"
      responses:
        "200":
          description: Question received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"

  /questions:
    get:
      operationId: getQuestions
      tags: [questions]
      summary: Get questions for user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"

  /questions/{id}/answer:
    post:
      operationId: answerQuestion
      tags: [questions]
      summary: Answer a question and forward to webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerRequest"
      responses:
        "200":
          description: Question answered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"

  /ml-oauth/connect:
    get:
      operationId: getMlOAuthUrl
      tags: [ml-oauth]
      summary: Get Mercado Livre OAuth URL
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OAuth URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthUrlResponse"

  /ml-oauth/callback:
    get:
      operationId: mlOAuthCallback
      tags: [ml-oauth]
      summary: Handle Mercado Livre OAuth callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MlConnectionStatus"

  /ml-oauth/status:
    get:
      operationId: getMlConnectionStatus
      tags: [ml-oauth]
      summary: Get Mercado Livre connection status
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MlConnectionStatus"

  /ml-oauth/disconnect:
    delete:
      operationId: disconnectMl
      tags: [ml-oauth]
      summary: Disconnect Mercado Livre
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disconnected

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          minLength: 6
      required:
        - email
        - name
        - password

    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
        token:
          type: string
      required:
        - user
        - token

    UserInfo:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
      required:
        - id
        - email
        - name

    Product:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        sku:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
        stockQuantity:
          type: integer
        category:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        mercadoLivreId:
          type: string
          nullable: true
        shopeeId:
          type: string
          nullable: true
        amazonId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - sku
        - price
        - stockQuantity
        - createdAt
        - updatedAt

    CreateProductRequest:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        sku:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
        stockQuantity:
          type: integer
        category:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required:
        - userId
        - name
        - sku
        - price
        - stockQuantity

    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
        stockQuantity:
          type: integer
        category:
          type: string
          nullable: true
      required:
        - name
        - sku
        - price
        - stockQuantity

    DashboardStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalStock:
          type: integer
        lowStockCount:
          type: integer
        syncedPlatforms:
          type: integer
        recentSyncs:
          type: array
          items:
            $ref: "#/components/schemas/SyncLog"
      required:
        - totalProducts
        - totalStock
        - lowStockCount
        - syncedPlatforms
        - recentSyncs

    SyncRequest:
      type: object
      properties:
        userId:
          type: integer
        platform:
          type: string
          enum: [mercadolivre, shopee, amazon]
      required:
        - userId
        - platform

    SyncResult:
      type: object
      properties:
        platform:
          type: string
        status:
          type: string
        message:
          type: string
        syncedAt:
          type: string
          format: date-time
      required:
        - platform
        - status
        - message
        - syncedAt

    SyncLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        platform:
          type: string
        status:
          type: string
        message:
          type: string
          nullable: true
        syncedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - platform
        - status
        - syncedAt

    N8nQuestionPayload:
      type: object
      properties:
        pergunta:
          type: string
        resposta:
          type: string
          nullable: true
        status:
          type: string
        item_id:
          type: string
        thumbnail:
          type: string
          nullable: true
        permalink:
          type: string
          nullable: true
        loja:
          type: string
          nullable: true
        ml_question_id:
          type: string
        userId:
          type: integer
      required:
        - pergunta
        - status
        - item_id
        - ml_question_id

    Question:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        mlQuestionId:
          type: string
        pergunta:
          type: string
        resposta:
          type: string
          nullable: true
        status:
          type: string
        itemId:
          type: string
        thumbnail:
          type: string
          nullable: true
        permalink:
          type: string
          nullable: true
        loja:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        answeredAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - userId
        - mlQuestionId
        - pergunta
        - status
        - itemId
        - createdAt

    AnswerRequest:
      type: object
      properties:
        resposta:
          type: string
        userId:
          type: integer
      required:
        - resposta
        - userId

    OAuthUrlResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url

    MlConnectionStatus:
      type: object
      properties:
        connected:
          type: boolean
        nickname:
          type: string
          nullable: true
        mlUserId:
          type: string
          nullable: true
      required:
        - connected