# lib/api-zod/src/generated/types/answerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hub Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerRequest {
  resposta: string;
  userId: number;
}