# lib/api-zod/src/generated/types/mlConnectionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hub Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface MlConnectionStatus {
  connected: boolean;
  nickname?: string | null;
  mlUserId?: string | null;
}